支持素材路径 及 参数调节优化

+++ QVEditorKit/FrameWork/XYCommonEngineKit.framework/Versions/A/Headers/XYEngineWorkspaceConfiguration.h
@@ -23,7 +23,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, copy) NSString * _Nullable effectEffectMaskRadialPath; //径向蒙版
 @property (nonatomic, copy) NSString * _Nullable effectEffectMaskRectanglePath; //矩形蒙版
 
-@property (nonatomic, assign) long long adjustEffectId;//
+@property (nonatomic, assign) UInt64 adjustEffectId;//
 @property (nonatomic, copy) NSString * _Nullable adjustEffectPath; //参数调节模板路径
 
 @property (nonatomic, assign) CGFloat outPutResolutionWidth;
